Add label sanity check to entry formatting
Before returning a formatted entry check if the current entry options
carry a label. If they don't they are probably not an entry that should
be displayed since we don't know how to in the completion menu.
Instead they are in all probability for example a metadata entry
provided through JabRef.